Move project label state to projects slice
