+++ src/features/filters/utils.js
@@ -30,14 +30,7 @@ const updateAvailCamFilters = (state, camConfigs) => {
 
 const updateAvailLabelFilters = (state, labels) => {
   state.availFilters.labels.options =  labels;
-  const noneFound = (labels.length === 0);
-  state.availFilters.labels.loadingState = {
-    isLoading: false,
-    operation: null,
-    errors: null,
-    noneFound,
-  };
-}
+};
 
 export {
   normalizeFilters,
