+++ src/features/projects/projectsSlice.js
@@ -54,6 +54,11 @@ const initialState = {
       errors: null,
       progress: 0
     },
+    projectLabels: {
+      isLoading: false,
+      operation: null,
+      errors: null,
+    },
   },
   unsavedViewChanges: false,
   modalOpen: false,
@@ -163,7 +168,7 @@ export const projectsSlice = createSlice({
 
 
     /* 
-     * Views CRUD 
+     * Views CRUD
      */
 
     editViewStart: (state) => { 
@@ -209,7 +214,7 @@ export const projectsSlice = createSlice({
     },
 
     /* 
-     * Automation Rules CRUD 
+     * Automation Rules CRUD
      */
 
     updateAutomationRulesStart: (state) => {
@@ -271,20 +276,22 @@ export const projectsSlice = createSlice({
       state.loadingStates.deployments.errors.splice(index, 1);
     },
 
-    /* fetch model source records */
+    /* 
+     * fetch model source records
+     */
 
     getModelsStart: (state) => {
-      const ls = { isLoading: true, operation: 'fetching', errors: null };  
+      const ls = { isLoading: true, operation: 'fetching', errors: null };
       state.loadingStates.deployments = ls;
     },
 
     getModelsFailure: (state, { payload }) => {
-      const ls = { isLoading: false, operation: null, errors: payload };  
+      const ls = { isLoading: false, operation: null, errors: payload };
       state.loadingStates.deployments = ls;
     },
 
     getModelsSuccess: (state, { payload }) => {
-      const ls = { isLoading: false, operation: null, errors: null };  
+      const ls = { isLoading: false, operation: null, errors: null };
       state.loadingStates.deployments = ls;
 
       const proj = state.projects.find((p) => p._id === payload.projId);
@@ -295,7 +302,6 @@ export const projectsSlice = createSlice({
     },
 
     getModelOptionsStart: (state) => {
-      console.log('getModelOptionsStart')
       const ls = { isLoading: true, operation: 'fetching', errors: null };
       state.loadingStates.modelOptions = ls;
     },
@@ -306,10 +312,76 @@ export const projectsSlice = createSlice({
     },
 
     getModelOptionsSuccess: (state, { payload }) => {
-      console.log('getModelOptionsSuccess')
       const ls = { isLoading: false, operation: null, errors: null };
       state.loadingStates.modelOptions = ls;
       state.modelOptions = payload;
+    },
+
+    dismissModelsError: (state, { payload }) => {
+      const index = payload;
+      state.loadingStates.models.errors.splice(index, 1);
+    },
+
+    /* 
+     * Project Labels CRUD
+     */
+
+    createProjectLabelStart: (state) => {
+      const ls = { isLoading: true, operation: 'creating', errors: null };
+      state.loadingStates.projectLabels.loadingState = ls;
+    },
+
+    createProjectLabelSuccess: (state, { payload }) => {
+      const ls = {
+        isLoading: false,
+        operation: null,
+        errors: null
+      };
+      state.loadingStates.projectLabels.loadingState = ls;
+
+      const proj = state.projects.find((p) => p._id === payload.projId);
+      proj.labels = [
+        ...proj.labels,
+        payload.label
+      ];
+    },
+
+    createProjectLabelFailure: (state, { payload }) => {
+      const ls = { isLoading: false, operation: null, errors: payload, stateMsg: null };
+      state.loadingStates.projectLabels.loadingState = ls;
+    },
+
+    updateProjectLabelStart: (state) => {
+      const ls = { isLoading: true, operation: 'updating', errors: null };
+      state.loadingStates.projectLabels.loadingState = ls;
+    },
+
+    updateProjectLabelSuccess: (state, { payload }) => {
+      const ls = {
+        isLoading: false,
+        operation: null,
+        errors: null
+      };
+      state.loadingStates.projectLabels.loadingState = ls;
+
+      const proj = state.projects.find((p) => p._id === payload.projId);
+      proj.labels = proj.labels.map((label) => {
+        if (label._id === payload.label._id) {
+          return payload.label;
+        } else {
+          return label;
+        }
+      });
+    },
+
+    updateProjectLabelFailure: (state, { payload }) => {
+      const ls = { isLoading: false, operation: null, errors: payload, stateMsg: null };
+      state.availFilters.labels.loadingState = ls;
+    },
+
+    dismissManageLabelsError: (state, { payload }) => {
+      const index = payload;
+      state.availFilters.labels.loadingState.errors.splice(index, 1);
     },
 
     setModalOpen: (state, { payload }) => {
@@ -318,11 +390,6 @@ export const projectsSlice = createSlice({
 
     setModalContent: (state, { payload }) => {
       state.modalContent = payload;
-    },
-
-    dismissModelsError: (state, { payload }) => {
-      const index = payload;
-      state.loadingStates.models.errors.splice(index, 1);
     },
   },
 
@@ -380,10 +447,18 @@ export const {
   getModelsStart,
   getModelsFailure,
   getModelsSuccess,
-  dismissModelsError,
   getModelOptionsStart,
   getModelOptionsFailure,
   getModelOptionsSuccess,
+  dismissModelsError,
+
+  createProjectLabelStart,
+  createProjectLabelSuccess,
+  createProjectLabelFailure,
+  updateProjectLabelStart,
+  updateProjectLabelSuccess,
+  updateProjectLabelFailure,
+  dismissManageLabelsError,
 
   setModalOpen,
   setModalContent,
@@ -608,7 +683,58 @@ export const fetchModelOptions = () => {
       dispatch(getModelOptionsFailure(err));
     }
   };
-}
+};
+
+// Project Labels thunks
+export const createProjectLabel = (payload) => {
+  return async (dispatch, getState) => {
+    try {
+      dispatch(createProjectLabelStart());
+      const currentUser = await Auth.currentAuthenticatedUser();
+      const token = currentUser.getSignInUserSession().getIdToken().getJwtToken();
+      const projects = getState().projects.projects;
+      const selectedProj = projects.find((proj) => proj.selected);
+      const projId = selectedProj._id;
+
+      if (token && selectedProj) {
+        const res = await call({
+          projId,
+          request: 'createProjectLabel',
+          input: payload
+        });
+        dispatch(createProjectLabelSuccess({ projId, label: res.createProjectLabel.label}));
+      }
+    } catch (err) {
+      console.log(`error attempting to create label: `, err);
+      dispatch(createProjectLabelFailure(err));
+    }
+  };
+};
+
+export const updateProjectLabel = (payload) => {
+  return async (dispatch, getState) => {
+    try {
+      dispatch(updateProjectLabelStart());
+      const currentUser = await Auth.currentAuthenticatedUser();
+      const token = currentUser.getSignInUserSession().getIdToken().getJwtToken();
+      const projects = getState().projects.projects;
+      const selectedProj = projects.find((proj) => proj.selected);
+      const projId = selectedProj._id;
+
+      if (token && selectedProj) {
+        const res = await call({
+          projId,
+          request: 'updateProjectLabel',
+          input: payload
+        });
+        dispatch(updateProjectLabelSuccess({ projId, label: res.updateProjectLabel.label }));
+      }
+    } catch (err) {
+      console.log(`error attempting to update label: `, err);
+      dispatch(updateProjectLabelFailure(err));
+    }
+  };
+};
 
 // Selectors
 export const selectProjects = state => state.projects.projects;
@@ -643,6 +769,8 @@ export const selectCreateProjectsErrors = state => state.projects.loadingStates.
 export const selectCreateProjectLoading = state => state.projects.loadingStates.createProject.isLoading;
 export const selectModelOptions = state => state.projects.modelOptions;
 export const selectModelOptionsLoading = state => state.projects.loadingStates.modelOptions.isLoading;
+export const selectProjectLabelsLoading = state => state.projects.loadingStates.projectLabels;
+export const selectManageLabelsErrors = state => state.projects.loadingStates.projectLabels.errors;
 
 
 export default projectsSlice.reducer;
